requirements.txt
playwright
httpx[http2]
asyncpg
python-dotenv
rich
tqdm

# ── Migrations ─────────────────────────────────────────────────────────────────
alembic==1.16.5
sqlalchemy
psycopg2-binary

# ── API ────────────────────────────────────────────────────────────────────────
fastapi==0.128.8
uvicorn[standard]

# ── AWS ────────────────────────────────────────────────────────────────────────
boto3

# ── Indexer ─────────────────────────────────────────────────────────────────────
pdfminer.six
